Move handleSQLError to tx file
